+++ test/index.js
@@ -1,5 +1,5 @@
 describe("multi-rpc-common", function () {
     require("./Serializer");
     require("./Transport");
-    require("./PersistantTransport");
+    require("./PersistentTransport");
 });