+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## [2.0.0] - 2018-12-20
+### Changes
+- Corrects spelling.
+
 ## [1.0.8] - 2018-12-20
 ### Changes
-- Adds errors from PersistantTransport.
+- Adds errors from PersistentTransport.
 
 ## [1.0.5] - 2018-12-20
 ### Changes
@@ -14,7 +18,7 @@
 
 ## [1.0.3] - 2018-12-19
 ### Added
-- Adds persistant transport
+- Adds Persistent transport
 
 ## [1.0.2] - 2018-12-18
 ### Added